fix(editor): move SlateOverlay from core/hooks into editor-ui
